Add support for LFS_MULTIVERSION.
LittleFS 2.6.0 changed the on-disk minor version from lfs2.0 to 2.1.
The first write updates an existing file system from 2.0 to 2.1.
An updated file system can not be mounted with older firmwares, see
for example https://github.com/littlefs-project/littlefs/issues/825

LittleFS 2.7.0 added the LFS_MULTIVERSION define and the disk_version
config to fix this.

This adds config entries to set LFS_MULTIVERSION and select the
disk version to write.